requirements.txt
farm-haystack[faiss]